Minor refactoring - rename
