+++ src/utils/uuid-validator.js
@@ -1,2 +1,2 @@
-module.exports.isValidId = (uuid) =>
-  (/^[0-9a-f]{8}-[0-9a-f]{4}-[0-5][0-9a-f]{3}-[089ab][0-9a-f]{3}-[0-9a-f]{12}$/i).test(uuid);
+module.exports.isValidId = (id) =>
+  id.match(/^[0-9a-f]{8}-[0-9a-f]{4}-[0-5][0-9a-f]{3}-[089ab][0-9a-f]{3}-[0-9a-f]{12}$/i); // uuid
